Beginnings of a tiny inetd
We'll be running a socket listen loop and forking to run an inetd-aware
program.
The intention is to be suitable for running in an interactive testing
session.  There are no plans to be suitable for serious use as a daemon.

This commit will not compile!  (At least, I don't think so - I
haven't tried it.)
Socket code still has to be written, and the fork-and-exec code
should not be trusted until it's tested (it's my first fork-and-exec,
ever).